Extract record search boundary
